feat: implement package management service with DTOs and validation

+++ Tradof.Data/Interfaces/IGeneralRepository.cs
@@ -5,10 +5,10 @@ namespace Tradof.Data.Interfaces
     public interface IGeneralRepository<T> where T : class
     {
         Task<IEnumerable<T>> GetAllAsync();
-        Task<T?> GetByIdAsync(int id);
+        Task<T?> GetByIdAsync(long id);
         Task<IEnumerable<T>> FindAsync(Expression<Func<T, bool>> predicate);
         Task AddAsync(T entity);
         Task UpdateAsync(T entity);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
